Change data types for less memory usage and faster controls

+++ src/correlator.cuh
@@ -8,20 +8,20 @@
 namespace MultiTau {
 
 
-    __inline__ __device__ size_t
-    insert_until_bin(const size_t instants, const size_t num_bins);
+    __inline__ __device__ int
+    insert_until_bin(const int instants, const int num_bins);
 
     template <typename T>
     __global__ void 
     correlate ( T * new_values, 
-                const size_t timepoints, 
-                size_t instants_processed, 
+                const int timepoints, 
+                int instants_processed, 
                 T * shift_register, 
                 int * shift_positions, 
                 T * accumulators, 
                 T * correlation,  
-                const size_t num_bins,
-                const size_t num_sensors
+                const int num_bins,
+                const int num_sensors
                 );
 
 }
@@ -30,37 +30,37 @@ template<typename T>
 class Correlator {
 
     public:
-        Correlator(const size_t num_bins, const size_t bin_size, const size_t num_sensors, const size_t packet_size, const int device = 0, const bool debug = false);
+        Correlator(const int num_bins, const int bin_size, const int num_sensors, const int packet_size, const int device = 0, const bool debug = false);
         ~Correlator();
 
         void alloc();
-        void correlate(const T * new_values, const size_t timepoints);
+        void correlate(const T * new_values, const int timepoints);
         void transfer();
-        T get(const size_t sensor, const size_t lag);
+        T get(const int sensor, const int lag);
         void reset();
 
     private:
         dim3 number_of_blocks;
         dim3 threads_per_block;
 
-        size_t num_bins;
-        size_t bin_size;
-        size_t num_sensors;
-        size_t num_sensors_per_block;
-        size_t packet_size;
+        int num_bins;
+        int bin_size;
+        int num_sensors;
+        int num_sensors_per_block;
+        int packet_size;
 
-        size_t shared_memory_per_block;
+        int shared_memory_per_block;
 
         uint32_t max_tau;
         uint32_t num_taus;
 
-        size_t instants_processed = 0;
+        int instants_processed = 0;
 
         bool debug;
         bool transfered;
 
         T * correlation = nullptr;
-        size_t * taus = nullptr;
+        int * taus = nullptr;
 
         // Device variables
         T * d_shift_register = nullptr;
